+++ config/variables.py
@@ -21,6 +21,6 @@ ALERTING_SEVERITIES = ['medium', 'high', 'critical']
 ROOT_PATH = Path(__file__).parent.parent
 FINDING_RULES_PATH = ROOT_PATH / 'config' / 'rules.yaml'
 
-META_TYPES = ['EC2', 'ELBV2', 'IAM', 'RDS', 'S3']
+META_TYPES = ['APIGW', 'CLOUDFRONT', 'EC2', 'ELBV2', 'IAM', 'RDS', 'S3']
 
 ACTION_PASSLIST = ['ec2messages', 'logs', 'ssm', 'ssmmessages']