+++ libs/asset_type_iam_group.py
@@ -22,7 +22,7 @@ class IAMGroup(AssetType):
     IAMGroup Asset Type
     """
     def __init__(self, name: str):
-        super().__init__(name)
+        super().__init__('IAM roles', name)
         self.list = []
 
     def audit(self, security_config):